Move classes out of Keppel.kt
