+++ Android/app/src/test/java/uk/ac/lshtm/keppel/android/settings/DataStoreActivityTest.kt
@@ -19,7 +19,7 @@ import uk.ac.lshtm.keppel.android.R
 
 @RunWith(RobolectricTestRunner::class)
 @LooperMode(PAUSED)
-class SettingsActivityTest {
+class DataStoreActivityTest {
 
     @get:Rule
     val rule = ActivityScenarioRule(SettingsActivity::class.java)
